requirements.txt
fastapi~=0.115.12
uvicorn~=0.33.0
httpx~=0.27.0

google-api-python-client~=2.66.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.1

Jinja2>=3.0,<4.0
python-multipart~=0.0.6

pillow~=11.2.0
opencv-python~=4.11.0.86
scikit-image~=0.25.2
numpy>=1.24,<1.28
geopy~=2.4.1
itsdangerous==2.1.2
protobuf~=4.25.3
starlette~=0.46.1
tqdm~=4.67.0
google-cloud-bigquery>=3.0.0
aiohttp==3.9.3


requests~=2.32.3
python-dotenv~=1.1.0